+++ src/elements/deployed_addresses.rs
@@ -5,10 +5,8 @@ use crate::{
     utils::{
         address_verifier::AddressVerifier,
         facet_cut_set::{self, FacetCutSet, FacetInfo},
-        network_verifier::Bridgehub as BridgehubSol,
-        network_verifier::BridgehubInfo,
-    },
-    UpgradeOutput,
+        network_verifier::{Bridgehub as BridgehubSol, BridgehubInfo},
+    }, verifiers::VerificationResult, UpgradeOutput
 };
 use alloy::{
     primitives::{Address, U256},
@@ -214,8 +212,6 @@ pub struct DeployedAddresses {
     pub(crate) validator_timelock_addr: Address,
     pub(crate) l2_wrapped_base_token_store_addr: Address,
     pub(crate) l1_bytecodes_supplier_addr: Address,
-    pub(crate) rollup_l1_da_validator_addr: Address,
-    pub(crate) validium_l1_da_validator_addr: Address,
     pub(crate) l1_transitionary_owner: Address,
     pub(crate) l1_rollup_da_manager: Address,
     pub(crate) l1_governance_upgrade_timer: Address,
@@ -231,7 +227,6 @@ pub struct Bridges {
     pub l1_nullifier_implementation_addr: Address,
     pub l1_nullifier_proxy_addr: Address,
     pub erc20_bridge_implementation_addr: Address,
-    pub bridged_standard_erc20_impl: Address,
     pub bridged_token_beacon: Address,
 }
 
@@ -548,7 +543,7 @@ impl DeployedAddresses {
     }
 
     fn expected_previous_protocol_version() -> U256 {
-        U256::from(25) * U256::from(2).pow(U256::from(32))
+        U256::from(26) * U256::from(2).pow(U256::from(32))
     }
 
     async fn verify_ctm_deployment_tracker(
@@ -612,7 +607,7 @@ impl DeployedAddresses {
         let l1_asset_router_init_calldata =
             L1AssetRouter::initializeCall::new((config.deployer_addr,)).abi_encode();
 
-        /*result
+        result
             .expect_create2_params_proxy_with_bytecode(
                 verifiers,
                 &self.bridges.l1_asset_router_proxy_addr,
@@ -621,7 +616,7 @@ impl DeployedAddresses {
                 l1_asset_router_impl_constructor,
                 "l1-contracts/L1AssetRouter",
             )
-            .await;*/
+            .await;
 
         let provider = verifiers.network_verifier.get_l1_provider();
         let l1_asset_router = L1AssetRouter::new(self.bridges.l1_asset_router_proxy_addr, provider);
@@ -666,8 +661,8 @@ impl DeployedAddresses {
             bridgehub_info.bridgehub_addr,
             U256::from(config.era_chain_id),
             // TODO: for local setup, it is 0. For production should be era (for backwards compatibility).
-            Address::ZERO
-            //era_diamond_proxy,
+            //Address::ZERO
+            era_diamond_proxy,
         ))
         .abi_encode();
 
@@ -685,7 +680,6 @@ impl DeployedAddresses {
         config: &UpgradeOutput,
         verifiers: &crate::verifiers::Verifiers,
         result: &mut crate::verifiers::VerificationResult,
-        bridgehub_info: &BridgehubInfo,
     ) -> Result<()> {
         result.expect_create2_params(
             verifiers,
@@ -811,136 +805,10 @@ impl DeployedAddresses {
         Ok(())
     }
 
-    async fn verify_rollup_da_manager(
-        &self,
-        config: &UpgradeOutput,
-        verifiers: &crate::verifiers::Verifiers,
-        result: &mut crate::verifiers::VerificationResult,
-        _bridgehub_info: &BridgehubInfo,
-    ) -> Result<()> {
-        result.expect_create2_params(
-            verifiers,
-            &self.l1_rollup_da_manager,
-            Vec::new(),
-            "l1-contracts/RollupDAManager",
-        );
-
-        let provider = verifiers.network_verifier.get_l1_provider();
-        let rollup_da_manager = RollupDAManager::new(self.l1_rollup_da_manager, provider);
-        let is_rollup_pair_allowed = rollup_da_manager
-            .isPairAllowed(
-                self.rollup_l1_da_validator_addr,
-                config.contracts_config.expected_rollup_l2_da_validator,
-            )
-            .call()
-            .await?
-            ._0;
-        ensure!(
-            is_rollup_pair_allowed,
-            "Rollup pair not allowed in RollupDAManager"
-        );
-
-        let current_owner = rollup_da_manager.owner().call().await?.owner;
-        // replace with governance owner.
-        /*ensure!(
-            current_owner == self.l1_transitionary_owner,
-            "RollupDAManager owner mismatch"
-        );*/
-        Ok(())
-    }
-
-    async fn verify_transitionary_owner(
-        &self,
-        config: &UpgradeOutput,
-        verifiers: &crate::verifiers::Verifiers,
-        result: &mut crate::verifiers::VerificationResult,
-        _bridgehub_info: &BridgehubInfo,
-    ) -> Result<()> {
-        result.expect_create2_params(
-            verifiers,
-            &self.l1_transitionary_owner,
-            TransitionaryOwner::constructorCall::new((
-                config.protocol_upgrade_handler_proxy_address,
-            ))
-            .abi_encode(),
-            "l1-contracts/TransitionaryOwner",
-        );
-        Ok(())
-    }
-
-    async fn verify_bridged_token_beacon(
-        &self,
-        _config: &UpgradeOutput,
-        verifiers: &crate::verifiers::Verifiers,
-        result: &mut crate::verifiers::VerificationResult,
-        _bridgehub_info: &BridgehubInfo,
-    ) -> Result<()> {
-        result.expect_create2_params(
-            verifiers,
-            &self.bridges.bridged_token_beacon,
-            BridgedTokenBeacon::constructorCall::new((self.bridges.bridged_standard_erc20_impl,))
-                .abi_encode(),
-            "l1-contracts/UpgradeableBeacon",
-        );
-        Ok(())
-    }
-
-    async fn verify_message_root(
-        &self,
-        verifiers: &crate::verifiers::Verifiers,
-        result: &mut crate::verifiers::VerificationResult,
-        bridgehub_info: &BridgehubInfo,
-    ) -> Result<()> {
-        let message_root_impl_constructor =
-            MessageRoot::constructorCall::new((bridgehub_info.bridgehub_addr,)).abi_encode();
-        let message_root_init_calldata = MessageRoot::initializeCall::new(()).abi_encode();
-
-        result
-            .expect_create2_params_proxy_with_bytecode(
-                verifiers,
-                &self.bridgehub.message_root_proxy_addr,
-                message_root_init_calldata,
-                bridgehub_info.transparent_proxy_admin,
-                message_root_impl_constructor,
-                "l1-contracts/MessageRoot",
-            )
-            .await;
-        Ok(())
-    }
-
-    async fn verify_governance_upgrade_timer(
-        &self,
-        config: &UpgradeOutput,
-        verifiers: &crate::verifiers::Verifiers,
-        result: &mut crate::verifiers::VerificationResult,
-        bridgehub_info: &BridgehubInfo,
-    ) -> Result<()> {
-        let initial_delay = if config.era_chain_id == 300 || config.l1_chain_id == 1 {
-            2 * 7 * 24 * 3600
-        } else {
-            24 * 3600
-        };
-        const MAX_INITIAL_DELAY: u32 = 1209600;
-        let expected_constructor_params = GovernanceUpgradeTimer::constructorCall::new((
-            U256::from(initial_delay),
-            U256::from(MAX_INITIAL_DELAY),
-            config.protocol_upgrade_handler_proxy_address,
-            bridgehub_info.ecosystem_admin,
-        ))
-        .abi_encode();
-
-        result.expect_create2_params(
-            verifiers,
-            &self.l1_governance_upgrade_timer,
-            expected_constructor_params,
-            "l1-contracts/GovernanceUpgradeTimer",
-        );
-        Ok(())
-    }
-
     pub async fn get_expected_facet_cuts(
         &self,
         verifiers: &crate::verifiers::Verifiers,
+        result: &mut VerificationResult,
     ) -> anyhow::Result<(FacetCutSet, FacetCutSet)> {
         let bridgehub_addr = verifiers.bridgehub_address;
         let bridgehub_info = verifiers
@@ -976,6 +844,10 @@ impl DeployedAddresses {
                 .get_code_at(address)
                 .await
                 .context(format!("Failed to retrieve the bytecode for {}", address))?;
+
+            if bytecode.len() == 0 {
+                result.report_warn(&format!("No bytecode for facet {}", facet.name));
+            }
             let info: Vec<_> =
                 evmole::contract_info(evmole::ContractInfoArgs::new(&bytecode.0).with_selectors())
                     .functions
@@ -996,76 +868,19 @@ impl DeployedAddresses {
         Ok((facets_to_remove, facets_to_add))
     }
 
-    pub async fn verify_protocol_upgrade_handler_impl(
-        &self,
-        config: &UpgradeOutput,
-        verifiers: &crate::verifiers::Verifiers,
-        result: &mut crate::verifiers::VerificationResult,
-        bridgehub_info: &BridgehubInfo,
-    ) -> anyhow::Result<()> {
-        let protocol_upgrade_handler_impl = ProtocolUpgradeHandler::new(
-            config.protocol_upgrade_handler_impl_address,
-            verifiers.network_verifier.get_l1_provider(),
-        );
-
-        let bridgehub = protocol_upgrade_handler_impl
-            .BRIDGE_HUB()
-            .call()
-            .await
-            .unwrap()
-            .BRIDGE_HUB;
-        let chain_type_manager = protocol_upgrade_handler_impl
-            .CHAIN_TYPE_MANAGER()
-            .call()
-            .await
-            .unwrap()
-            .CHAIN_TYPE_MANAGER;
-        let era = protocol_upgrade_handler_impl
-            .ZKSYNC_ERA()
-            .call()
-            .await
-            .unwrap()
-            .ZKSYNC_ERA;
-        let l1_nullifier = protocol_upgrade_handler_impl
-            .L1_NULLIFIER()
-            .call()
-            .await
-            .unwrap()
-            .L1_NULLIFIER;
-        let l1_asset_router = protocol_upgrade_handler_impl
-            .L1_ASSET_ROUTER()
-            .call()
-            .await
-            .unwrap()
-            .L1_ASSET_ROUTER;
-        let l1_native_token_vault = protocol_upgrade_handler_impl
-            .L1_NATIVE_TOKEN_VAULT()
-            .call()
-            .await
-            .unwrap()
-            .L1_NATIVE_TOKEN_VAULT;
-
-        result.expect_address(verifiers, &bridgehub, "bridgehub_proxy");
-        result.expect_address(verifiers, &chain_type_manager, "state_transition_manager");
-        if bridgehub_info.era_address != era {
-            result.report_error(&format!(
-                "Incorrect era address. Expected {}, received: {}",
-                bridgehub_info.era_address, era
-            ));
-        }
-        result.expect_address(verifiers, &l1_nullifier, "l1_nullifier_proxy");
-        result.expect_address(verifiers, &l1_asset_router, "l1_asset_router_proxy");
-        result.expect_address(verifiers, &l1_native_token_vault, "native_token_vault");
-
-        Ok(())
-    }
-
     pub async fn verify(
         &self,
         config: &UpgradeOutput,
         verifiers: &crate::verifiers::Verifiers,
         result: &mut crate::verifiers::VerificationResult,
     ) -> anyhow::Result<()> {
+        // Here we should verify all the addresses that we're deploying in a given upgrade.
+        // In case of v27, they are:
+        // * stm / ctm
+        // * bridgehub
+        // * l1 nullifier
+        // * l1 asset router
+        // * l1 native token vault
         let bridgehub_addr = verifiers.bridgehub_address;
         let bridgehub_info = verifiers
             .network_verifier
@@ -1089,10 +904,9 @@ impl DeployedAddresses {
         self.verify_l1_nullifier(config, verifiers, result, &bridgehub_info)
             .await
             .context("l1 nullifier")?;
-        self.verify_l1_erc20_bridge(config, verifiers, result, &bridgehub_info)
+        self.verify_l1_erc20_bridge(config, verifiers, result)
             .await
             .context("l1_erc20_bridge")?;
-        println!("here");
         self.verify_bridgehub_impl(config, verifiers, result)
             .await?;
         self.verify_chain_type_manager(config, verifiers, result, &bridgehub_info)
@@ -1105,23 +919,13 @@ impl DeployedAddresses {
             .await?;
         self.verify_mailbox_facet(config, verifiers, result, &bridgehub_info)
             .await?;
-        self.verify_rollup_da_manager(config, verifiers, result, &bridgehub_info)
-            .await?;
-        /*self.verify_transitionary_owner(config, verifiers, result, &bridgehub_info)
-            .await?;*/
-        /*self.verify_bridged_token_beacon(config, verifiers, result, &bridgehub_info)
-            .await?;*/
-        self.verify_message_root(verifiers, result, &bridgehub_info)
-            .await?;
-        /*self.verify_governance_upgrade_timer(config, verifiers, result, &bridgehub_info)
-            .await
-            .context("governance upgrade timer")?;*/
+
+
+
         self.verify_per_chain_info(config, verifiers, result, &bridgehub_info)
             .await
             .context("per chain info")?;
-        /*self.verify_protocol_upgrade_handler_impl(config, verifiers, result, &bridgehub_info)
-        .await
-        .context("protocol upgrade handler")?;*/
+
 
         // TODO: verify fflonk and plonk
 
@@ -1158,30 +962,7 @@ impl DeployedAddresses {
             Vec::new(),
             "l1-contracts/DiamondInit",
         );
-        result.expect_create2_params(
-            verifiers,
-            &self.l1_bytecodes_supplier_addr,
-            Vec::new(),
-            "l1-contracts/BytecodesSupplier",
-        );
-        result.expect_create2_params(
-            verifiers,
-            &self.rollup_l1_da_validator_addr,
-            Vec::new(),
-            "da-contracts/RollupL1DAValidator",
-        );
-        result.expect_create2_params(
-            verifiers,
-            &self.validium_l1_da_validator_addr,
-            Vec::new(),
-            "l1-contracts/ValidiumL1DAValidator",
-        );
-        result.expect_create2_params(
-            verifiers,
-            &self.bridges.bridged_standard_erc20_impl,
-            Vec::new(),
-            "l1-contracts/BridgedStandardERC20",
-        );
+ 
 
         result.report_ok("deployed addresses");
         Ok(())
