+++ src/elements/mod.rs
@@ -2,8 +2,7 @@ use alloy::primitives::{Address, FixedBytes, U256};
 use anyhow::Context;
 use call_list::CallList;
 use deployed_addresses::DeployedAddresses;
-use governance_stage1_calls::GovernanceStage1Calls;
-use governance_stage2_calls::GovernanceStage2Calls;
+use governance_stage_calls::{GovernanceStage0Calls, GovernanceStage1Calls, GovernanceStage2Calls};
 use initialize_data_new_chain::{FeeParams, PubdataPricingMode};
 use protocol_version::ProtocolVersion;
 use serde::Deserialize;
@@ -19,8 +18,7 @@ pub mod call_list;
 pub mod deployed_addresses;
 pub mod fixed_force_deployment;
 pub mod force_deployment;
-pub mod governance_stage1_calls;
-pub mod governance_stage2_calls;
+pub mod governance_stage_calls;
 pub mod initialize_data_new_chain;
 pub mod protocol_version;
 pub mod set_new_version_upgrade;
@@ -38,7 +36,6 @@ pub struct UpgradeOutput {
     pub(crate) l1_chain_id: u64,
 
     pub(crate) protocol_upgrade_handler_proxy_address: Address,
-    pub(crate) protocol_upgrade_handler_impl_address: Address,
 
     #[serde(rename = "contracts_newConfig")]
     pub(crate) contracts_config: ContractsConfig,
@@ -49,6 +46,7 @@ pub struct UpgradeOutput {
 
 #[derive(Debug, Deserialize)]
 pub struct GovernanceCalls {
+    pub(crate) governance_stage0_calls: String,
     pub(crate) governance_stage1_calls: String,
     pub(crate) governance_stage2_calls: String,
 }
@@ -217,7 +215,7 @@ impl UpgradeOutput {
             .context("checking deployed addresses")?;
         let (facets_to_remove, facets_to_add) = self
             .deployed_addresses
-            .get_expected_facet_cuts(verifiers)
+            .get_expected_facet_cuts(verifiers, result)
             .await
             .context("checking facets")?;
 
@@ -225,27 +223,25 @@ impl UpgradeOutput {
             .expect_deployed_bytecode(verifiers, &self.create2_factory_addr, "Create2Factory")
             .await;
 
+        let stage0 = GovernanceStage0Calls {
+            calls: CallList::parse(&self.governance_calls.governance_stage0_calls),
+        };
+
+        stage0
+            .verify(
+                verifiers,
+                result,
+            )
+            .await
+            .context("stage0")?;
+
         let stage1 = GovernanceStage1Calls {
             calls: CallList::parse(&self.governance_calls.governance_stage1_calls),
         };
 
         let expected_upgrade_facets = facets_to_remove.merge(facets_to_add.clone()).clone();
-
-        stage1
-            .verify(
-                &self.deployed_addresses,
-                verifiers,
-                result,
-                expected_upgrade_facets.clone(),
-                &self.chain_upgrade_diamond_cut,
-            )
-            .await
-            .context("stage1")?;
-
-        let stage2 = GovernanceStage2Calls {
-            calls: CallList::parse(&self.governance_calls.governance_stage2_calls),
-        };
-        let (expected_chain_creation_data, expected_force_deployments) = stage2
+        
+        let (expected_chain_creation_data, expected_force_deployments) = stage1
             .verify(
                 verifiers,
                 result,
@@ -255,7 +251,19 @@ impl UpgradeOutput {
                 &self.chain_upgrade_diamond_cut,
             )
             .await
-            .context("stage2")?;
+            .context("stage1")?;
+
+        let stage2 = GovernanceStage2Calls {
+            calls: CallList::parse(&self.governance_calls.governance_stage2_calls),
+        };
+
+        stage2
+        .verify(
+            verifiers,
+            result,
+        )
+        .await
+        .context("stage2")?;
 
         self.contracts_config
             .verify(
